Refactor task management into TaskList class
